condense invoice request and fix make payment
