--- src/Message/FetchSubscriptionInvoiceNumbersRequest.php
@@ -1,158 +1,0 @@
-<?php
-
-namespace Omnipay\Vindicia\Message;
-
-use Omnipay\Common\Exception\InvalidRequestException;
-
-/**
- * Fetch a Vindicia subscription object.
- *
- * Parameters:
- * - subscriptionId: Your identifier for the subscription to be fetched. Either subscriptionId
- * or subscriptionReference is required.
- * - subscriptionReference: The gateway's identifier for the subscription to be fetched. Either
- * subscriptionId or subscriptionReference is required.
- *
- * Example:
- * <code>
- *   // set up the gateway
- *   $gateway = \Omnipay\Omnipay::create('Vindicia');
- *   $gateway->setUsername('your_username');
- *   $gateway->setPassword('y0ur_p4ssw0rd');
- *   $gateway->setTestMode(false);
- *
- *   // create a plan to govern the behavior of the subscription (eg billing frequency)
- *   $planResponse = $gateway->createPlan(array(
- *       'planId' => '123456789', // you choose this
- *       'interval' => 'week',
- *       'intervalCount' => 2,
- *       'prices' => array(
- *           'USD' => '9.99',
- *           'CAD' => '12.99'
- *       )
- *   ))->send();
- *
- *   if ($planResponse->isSuccessful()) {
- *       echo "Plan id: " . $planResponse->getPlanId() . PHP_EOL;
- *       echo "Plan reference: " . $planResponse->getPlanReference() . PHP_EOL;
- *   } else {
- *       // error handling
- *   }
- *
- *   // create a product using that plan for the user to subscribe to
- *   $productResponse = $gateway->createProduct(array(
- *       'productId' => '123456789', // you choose this
- *       'planId' => $planResponse->getPlanId(), // assign the plan you created above to it
- *       'statementDescriptor' => 'Statement descriptor',
- *       'duplicateBehavior' => CreateProductRequest::BEHAVIOR_SUCCEED_IGNORE,
- *       'prices' => array(
- *           'USD' => '9.99',
- *           'CAD' => '12.99'
- *       )
- *   ))->send();
- *
- *   if ($productResponse->isSuccessful()) {
- *       echo "Product id: " . $productResponse->getProductId() . PHP_EOL;
- *       echo "Product reference: " . $productResponse->getProductReference() . PHP_EOL;
- *   } else {
- *       // error handling
- *   }
- *
- *   // finally we can create the subscription!
- *   $subscriptionResponse = $gateway->createSubscription(array(
- *       'currency' => 'GBP',
- *       'customerId' => '123456', // will be created if it doesn't already exist
- *       'card' => array(
- *           'number' => '5555555555554444',
- *           'expiryMonth' => '01',
- *           'expiryYear' => '2020',
- *           'cvv' => '123',
- *           'postcode' => '12345'
- *       ),
- *       'paymentMethodId' => 'cc-123456', // this ID will be assigned to the card
- *       'subscriptionId' => '111111', // you choose this
- *       'productId' => $productResponse->getProductId(),
- *       'planId' => $planResponse->getPlanId() // not necessary since it's already on the product
- *   ))->send();
- *
- *   if ($subscriptionResponse->isSuccessful()) {
- *       echo "Subscription id: " . $subscriptionResponse->getSubscriptionId() . PHP_EOL;
- *       echo "Subscription reference: " . $subscriptionResponse->getSubscriptionReference() . PHP_EOL;
- *   } else {
- *       // error handling
- *   }
- *
- *   $fetchResponse = $gateway->fetchSubscription(array(
- *       'subscriptionId' => $subscriptionResponse->getSubscriptionId() // could also do by reference
- *   ))->send();
- *
- *   if ($fetchResponse->isSuccessful()) {
- *       var_dump($fetchResponse->getSubscription());
- *   }
- */
-class FetchSubscriptionInvoiceNumbersRequest extends AbstractRequest
-{
-    /**
-     * Returns the invoice state, one of 'Open', 'Due', 'Paid', 'Overdue', or 'WrittenOff'
-     *
-     * @return null|string
-     */
-    public function getInvoiceState()
-    {
-        return $this->getParameter('invoiceState');
-    }
-
-    /**
-     * Sets the invoice state
-     *
-     * @param string $value
-     * @return static
-     */
-    public function setInvoiceState($value)
-    {
-        return $this->setParameter('invoiceState', $value);
-    }
-
-    /**
-     * The name of the function to be called in Vindicia's API
-     *
-     * @return string
-     */
-    protected function getFunction()
-    {
-        return 'fetchInvoiceNumbers';
-    }
-
-    /**
-     * @return string
-     */
-    protected function getObject()
-    {
-        return self::$SUBSCRIPTION_OBJECT;
-    }
-
-    public function getData()
-    {
-        $subscriptionId = $this->getSubscriptionId();
-        $subscriptionReference = $this->getSubscriptionReference();
-
-        if (!$subscriptionId && !$subscriptionReference) {
-            throw new InvalidRequestException(
-                'Either the subscriptionId or subscriptionReference parameter is required.'
-            );
-        }
-
-        $subscription = new stdClass();
-        $subscription->merchantAutoBillId = $subscriptionId;
-        $subscription->VID = $subscriptionReference;
-
-        $data = array(
-            'action' => $this->getFunction()
-        );
-
-        $data['autobill'] = $subscription;
-        $data['invoicestate'] = $this->getInvoiceState(); // not camel case as described in the doc
-
-        return $data;
-    }
-}
