+++ src/Message/MakePaymentRequest.php
@@ -5,116 +5,45 @@ namespace Omnipay\Vindicia\Message;
 use Omnipay\Common\Exception\InvalidRequestException;
 
 /**
- * Fetch a Vindicia subscription object.
+ * Make payment against an outstanding invoice amount.
  *
  * Parameters:
  * - subscriptionId: Your identifier for the subscription to be fetched. Either subscriptionId
  * or subscriptionReference is required.
  * - subscriptionReference: The gateway's identifier for the subscription to be fetched. Either
  * subscriptionId or subscriptionReference is required.
+ * - paymentMethodId: The identifier for the payment method used to make the payment. Either
+ * paymentMethodId or paymentMethodReference is required.
+ * - paymentMethodReference: The gateway's identifier (VID) for the payment method used to
+ * make the payment. Either paymentMethodId or paymentMethodReference is required.
+ * - amount: The amount of the outstanding invoice balance. That should be obtained from reading
+ * the invoice text to get the total amount due.
  *
  * Example:
  * <code>
- *   // set up the gateway
- *   $gateway = \Omnipay\Omnipay::create('Vindicia');
- *   $gateway->setUsername('your_username');
- *   $gateway->setPassword('y0ur_p4ssw0rd');
- *   $gateway->setTestMode(false);
+ *   // first create gateway, plan, product and subscription as described in FetchSubsciptionRequest
+ *   // then use subscriptionId returned from $subscriptionResponse->getSubscriptionId(),
+ *   // the payment method id from payment method record and the amount due
  *
- *   // create a plan to govern the behavior of the subscription (eg billing frequency)
- *   $planResponse = $gateway->createPlan(array(
- *       'planId' => '123456789', // you choose this
- *       'interval' => 'week',
- *       'intervalCount' => 2,
- *       'prices' => array(
- *           'USD' => '9.99',
- *           'CAD' => '12.99'
- *       )
+ *   // get $subscriptionResponse from createSubscription call
+ *
+ *   $makePaymentResponse = $gateway->makePayment(array(
+ *       'subscriptionId' => $subscriptionResponse->getSubscriptionId(), // could also do by reference
+ *       'paymentMethodId' => 'cc-123456' // could also do by reference (VID),
+ *       'amount' => 59.95
  *   ))->send();
  *
- *   if ($planResponse->isSuccessful()) {
- *       echo "Plan id: " . $planResponse->getPlanId() . PHP_EOL;
- *       echo "Plan reference: " . $planResponse->getPlanReference() . PHP_EOL;
- *   } else {
- *       // error handling
+ *   if ($makePaymentResponse->isSuccessful()) {
+ *       echo 'Payment summary " . $makePaymentResponse->getSummary() . PHP_EOL;';
+ *       vardump($makePaymentResponse->getTransaction());
  *   }
  *
- *   // create a product using that plan for the user to subscribe to
- *   $productResponse = $gateway->createProduct(array(
- *       'productId' => '123456789', // you choose this
- *       'planId' => $planResponse->getPlanId(), // assign the plan you created above to it
- *       'statementDescriptor' => 'Statement descriptor',
- *       'duplicateBehavior' => CreateProductRequest::BEHAVIOR_SUCCEED_IGNORE,
- *       'prices' => array(
- *           'USD' => '9.99',
- *           'CAD' => '12.99'
- *       )
- *   ))->send();
- *
- *   if ($productResponse->isSuccessful()) {
- *       echo "Product id: " . $productResponse->getProductId() . PHP_EOL;
- *       echo "Product reference: " . $productResponse->getProductReference() . PHP_EOL;
- *   } else {
- *       // error handling
- *   }
- *
- *   // finally we can create the subscription!
- *   $subscriptionResponse = $gateway->createSubscription(array(
- *       'currency' => 'GBP',
- *       'customerId' => '123456', // will be created if it doesn't already exist
- *       'card' => array(
- *           'number' => '5555555555554444',
- *           'expiryMonth' => '01',
- *           'expiryYear' => '2020',
- *           'cvv' => '123',
- *           'postcode' => '12345'
- *       ),
- *       'paymentMethodId' => 'cc-123456', // this ID will be assigned to the card
- *       'subscriptionId' => '111111', // you choose this
- *       'productId' => $productResponse->getProductId(),
- *       'planId' => $planResponse->getPlanId() // not necessary since it's already on the product
- *   ))->send();
- *
- *   if ($subscriptionResponse->isSuccessful()) {
- *       echo "Subscription id: " . $subscriptionResponse->getSubscriptionId() . PHP_EOL;
- *       echo "Subscription reference: " . $subscriptionResponse->getSubscriptionReference() . PHP_EOL;
- *   } else {
- *       // error handling
- *   }
- *
- *   $fetchResponse = $gateway->fetchSubscription(array(
- *       'subscriptionId' => $subscriptionResponse->getSubscriptionId() // could also do by reference
- *   ))->send();
- *
- *   if ($fetchResponse->isSuccessful()) {
- *       var_dump($fetchResponse->getSubscription());
- *   }
+ * </code>
  */
 class MakePaymentRequest extends AbstractRequest
 {
     /**
-     * Returns the payment method object for making autobill payment
-     *
-     * @return null|\Omnipay\Vindicia\PaymentMethod
-     */
-    public function getPaymentMethod()
-    {
-        return $this->getParameter('paymentMethod');
-    }
-
-    /**
-     * Sets the payment method
-     *
-     * @param string $value
-     * @return static
-     */
-    public function setPaymentMethod($value)
-    {
-        return $this->setParameter('paymentMethod', $value);
-    }
-
-    /**
-     * Returns the invoice id
+     * Returns the overdue amount
      *
      * @return null|string
      */
@@ -124,7 +53,7 @@ class MakePaymentRequest extends AbstractRequest
     }
 
     /**
-     * Sets the invoice id
+     * Sets the overdue amount
      *
      * @param string $value
      * @return static
@@ -132,27 +61,6 @@ class MakePaymentRequest extends AbstractRequest
     public function setAmount($value)
     {
         return $this->setParameter('amount', $value);
-    }
-
-        /**
-     * Returns the invoice id
-     *
-     * @return null|string
-     */
-    public function getInvoiceId()
-    {
-        return $this->getParameter('invoiceId');
-    }
-
-    /**
-     * Sets the invoice id
-     *
-     * @param string $value
-     * @return static
-     */
-    public function setInvoiceId($value)
-    {
-        return $this->setParameter('invoiceId', $value);
     }
 
     /**
@@ -177,11 +85,23 @@ class MakePaymentRequest extends AbstractRequest
     {
         $subscriptionId = $this->getSubscriptionId();
         $subscriptionReference = $this->getSubscriptionReference();
-
         if (!$subscriptionId && !$subscriptionReference) {
             throw new InvalidRequestException(
                 'Either the subscriptionId or subscriptionReference parameter is required.'
             );
+        }
+
+        $paymentMethodId = $this->getPaymentMethodId();
+        $paymentMethodReference = $this->getPaymentMethodReference();
+        if (!$paymentMethodId && !$paymentMethodReference) {
+            throw new InvalidRequestException(
+                'Either the $paymentMethodId or $paymentMethodReference parameter is required.'
+            );
+        }
+
+        $amount = $this->getAmount();
+        if (!$amount) {
+            throw new InvalidRequestException('$amount parameter is required.');
         }
 
         $subscription = new stdClass();
@@ -193,10 +113,10 @@ class MakePaymentRequest extends AbstractRequest
         );
 
         $data['autobill'] = $subscription;
-        $data['paymentMethod'] = $this->getPaymentMethod();
-        $data['amount'] = $this->getAmount();
+        $data['paymentMethod'] = $this->buildPaymentMethod();
+        $data['amount'] = $amount;
         $data['currency'] = null;
-        $data['invoiceId'] = $this->getInvoiceId();
+        $data['invoiceId'] = null;
         $data['overageDisposition'] = null;
         $data['note'] = 'Payment initiated by makePayment';
